BoardsConfigServiceTest test refactored and improved
