+++ cogboard-app/src/main/kotlin/com/cognifide/cogboard/storage/VolumeStorage.kt
@@ -6,19 +6,22 @@ import io.vertx.core.json.JsonObject
 import io.vertx.core.logging.Logger
 import io.vertx.core.logging.LoggerFactory
 import java.io.File
+import java.net.URL
 
 class VolumeStorage(
-    private val type: ConfigType,
-    private val configFile: String = type.configFilePath()
+        private val type: ConfigType,
+        private val configFile: String = type.configFilePath()
 ) : Storage {
 
     override fun loadConfig(): JsonObject {
-            val conf = File(configFile).readText()
-            return if (type.validate(conf)) JsonObject(conf)
-            else CogboardConstants.errorResponse("$type config not valid")
-        }
+        createIfDoesNotExist(configFile)
+        val conf = File(configFile).readText()
+        return if (type.validate(conf)) JsonObject(conf)
+        else CogboardConstants.errorResponse("$type config not valid")
+    }
 
     override fun saveConfig(configJson: JsonObject): Boolean {
+        createIfDoesNotExist(configFile)
         val conf = configJson.toString()
         if (type.validate(conf)) {
             File(configFile).writeText(conf)
@@ -27,6 +30,27 @@ class VolumeStorage(
 
         LOGGER.error("$ERROR_MESSAGE \nconfig:\n$configJson")
         return false
+    }
+
+    private fun createIfDoesNotExist(configPath: String) {
+        if (!File(configPath).exists()) {
+            createFile(configPath)
+        }
+    }
+
+    private fun createFile(configPath: String) {
+        val initFileName = getInitFileName(configPath)
+        val fileContent = getResource(initFileName).readText()
+        File(configPath).writeText(fileContent)
+        LOGGER.info("Configuration file created: $configPath")
+    }
+
+    private fun getInitFileName(configPath: String): String {
+        return "/initData/${configPath.substringAfterLast("/")}"
+    }
+
+    private fun getResource(initFileName: String): URL {
+        return VolumeStorageFactory::class.java.getResource(initFileName)
     }
 
     companion object {
